Add request_id of payload to be signed in notification + send data msg
Add request_id of the payload to be signed by the mobile app, to the
push notification.

Also, set empty title and body (in fact, an empty notification
attribute in the payload) to send the FCM message as a data
message (which is handled by the app even if the app is
running in background, on both ios and android) instead of as
 notification message (handled by the OS if app in background,
the OS's system tray makes a notification and only handles the
control to the app once the user clicks the notification).